[REFACTOR] Extract turn loop module
